Clean up Analysis
- fix icon
- remove unneeded classes
